requirements.txt
gradio>=4.19.2
replicate>=0.22.0
Pillow>=10.2.0
requests>=2.31.0
openai>=1.12.0
python-dotenv>=1.0.0
